test: relocate spec test
