+++ test/unit/modules/api.publish.spec.js
