+++ test/unit/modules/auth-utils.spec.js
