+++ test/unit/modules/auth.spec.js
