+++ test/unit/modules/cli.spec.js
