+++ test/unit/modules/config-utils.spec.js
