+++ test/unit/modules/config.spec.js
