+++ test/unit/modules/local-storage.spec.js
