+++ test/unit/modules/no_proxy.spec.js
