+++ test/unit/modules/notify-request.spec.js
