+++ test/unit/modules/notify.spec.js
