+++ test/unit/modules/parse_interval.spec.js
