+++ test/unit/modules/plugin_loader.spec.js
