+++ test/unit/modules/search.spec.js
