+++ test/unit/modules/st_merge.spec.js
