+++ test/unit/modules/storage-utils.spec.js
