+++ test/unit/modules/store.spec.js
