+++ test/unit/modules/uplink.auth.spec.js
