+++ test/unit/modules/validate_all.spec.js
